+++ hrms/src/main/java/kodlama/ht6/hrms/dataAccess/abstracts/JobDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import kodlama.ht6.hrms.entities.concretes.Job;
 
-public interface JobDao extends JpaRepository<Job, Integer>{
-	Job getByid(Integer id);
+public interface JobDao extends JpaRepository<Job, Short>{
+	Job getByid(Short id);
 }
